services/ml_service_image/requirements.txt
numpy
pydicom
scikit-image
torch==2.2.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
pillow